refactor(portfolio): extract usePortfolioOverview/useBenchmarks + sub-surfaces (FINLYNQ-118 phase 3)
Behaviour-preserving extraction of the portfolio god-component. page.tsx
1996 -> 593 lines (~70%). NO data-fetching library (FINLYNQ-115); the
bespoke fetch/useState/useEffect/finally pattern is kept verbatim.

- _types.ts: types + constants + clientCanonicalKey (verbatim).
- _hooks/use-portfolio.ts: usePortfolioOverview(currency) + useBenchmarks(period, enabled).
- _components/: portfolio-ui (tooltips/ChangeBadge/DayChange/skeleton), csv,
  allocation-charts, benchmark-chart, etf-xray-card, holdings-by-account, holdings-table.

The page keeps page-level view state + the derived memos (filteredHoldings /
holdingsByCanonicalKey / accountGroups) with their `(x ?? "").localeCompare`
null-guards intact, the hero/P&L/top-movers blocks, and the holding dialog.
