+++ src/app/(app)/portfolio/page.tsx
@@ -1,387 +1,41 @@
 "use client";
 
-import { useEffect, useState, useMemo } from "react";
+import { useState, useMemo } from "react";
 import Link from "next/link";
 import { Card, CardContent, CardHeader, CardTitle } from "@/components/ui/card";
-import { Table, TableBody, TableCell, TableHead, TableHeader, TableRow } from "@/components/ui/table";
-import { Badge } from "@/components/ui/badge";
 import { Button } from "@/components/ui/button";
 import { Dialog, DialogContent, DialogHeader, DialogTitle } from "@/components/ui/dialog";
 import {
-  PieChart, Pie, Cell, ResponsiveContainer, Tooltip,
-  LineChart, Line, XAxis, YAxis, Legend,
-} from "recharts";
-import {
-  TrendingUp, Wallet, BarChart3, Coins, ArrowUpRight, ArrowDownRight,
-  Globe2, Building2, ChevronDown, ChevronUp, ChevronLeft, ChevronRight, Layers, PieChart as PieChartIcon,
-  Briefcase, DollarSign, Flame, Snowflake, Search, Download, Plus,
+  TrendingUp, BarChart3, Coins, Briefcase, Plus, Flame, Snowflake,
 } from "lucide-react";
 import { formatCurrency } from "@/lib/currency";
-import { motion, AnimatePresence } from "framer-motion";
 import { useDevMode } from "@/hooks/use-dev-mode";
 import { useDisplayCurrency } from "@/components/currency-provider";
-import { isMetalCurrency } from "@/lib/fx/supported-currencies";
 import {
   HoldingEditForm,
   type HoldingEditFormHolding,
 } from "@/components/holdings/holding-edit-form";
-import { ColorDot, CspSafeColorBar } from "@/components/csp-safe-bar";
 import { PerformanceChart } from "@/components/portfolio/PerformanceChart";
 import { buttonVariants } from "@/components/ui/button";
 
-// Mirror of /api/portfolio/overview's `canonicalKey()`. Keep in sync with
-// the server-side function — both must produce the same key for a given
-// holding so the row-membership lookup matches the API's `byHolding` array.
-function clientCanonicalKey(h: { assetType: string; symbol: string | null; currency: string; name: string }): string {
-  const sym = h.symbol ? h.symbol.trim().toUpperCase() : "";
-  if (h.assetType === "crypto" && sym) return `crypto:${sym}`;
-  if ((h.assetType === "stock" || h.assetType === "etf") && sym) return `eq:${sym}`;
-  if (h.assetType === "cash") {
-    if (sym) {
-      if (isMetalCurrency(sym)) return `metal:${sym}`;
-      return `cash:${sym}`;
-    }
-    const cur = (h.currency || "").toUpperCase();
-    return `cash:${cur}`;
-  }
-  return `custom:${(h.name || "?").trim().toLowerCase()}`;
-}
-
-// ── Colors ──────────────────────────────────────────────────────────
-const PIE_COLORS = [
-  "#6366f1", "#06b6d4", "#10b981", "#f59e0b", "#f43f5e",
-  "#8b5cf6", "#14b8a6", "#84cc16", "#ec4899", "#f97316",
-];
-
-const ASSET_TYPE_CONFIG: Record<string, { label: string; color: string; icon: typeof Coins }> = {
-  etf: { label: "ETFs", color: "#6366f1", icon: Layers },
-  stock: { label: "Stocks", color: "#06b6d4", icon: BarChart3 },
-  crypto: { label: "Crypto", color: "#f59e0b", icon: Coins },
-  cash: { label: "Cash", color: "#10b981", icon: DollarSign },
-};
-
-const REGION_COLORS: Record<string, string> = {
-  US: "#6366f1", Canada: "#10b981", Europe: "#f59e0b", Japan: "#f43f5e",
-  Asia: "#8b5cf6", Emerging: "#06b6d4", Other: "#64748b",
-};
-
-const SECTOR_COLORS: Record<string, string> = {
-  Tech: "#6366f1", Healthcare: "#10b981", Financials: "#f59e0b",
-  Consumer: "#f43f5e", Industrials: "#06b6d4", Energy: "#8b5cf6",
-  Materials: "#14b8a6", Other: "#64748b",
-};
-
-// ── Types ───────────────────────────────────────────────────────────
-type AssetType = "etf" | "stock" | "crypto" | "cash";
-
-type EnrichedHolding = {
-  id: number;
-  accountId: number | null;
-  accountName: string;
-  name: string;
-  symbol: string | null;
-  currency: string;
-  assetType: AssetType;
-  price: number | null;
-  change: number | null;
-  changePct: number | null;
-  dayChangeDisplay: number | null;
-  quoteCurrency: string | null;
-  marketCap: number | null;
-  image: string | null;
-  quantity: number | null;
-  avgCostPerShare: number | null;
-  totalCostBasis: number | null;
-  lifetimeCostBasis: number | null;
-  marketValue: number | null;
-  marketValueDisplay: number | null;
-  unrealizedGain: number | null;
-  unrealizedGainPct: number | null;
-  unrealizedGainDisplay: number | null;
-  realizedGain: number | null;
-  dividendsReceived: number | null;
-  totalReturn: number | null;
-  totalReturnDisplay: number | null;
-  totalReturnPct: number | null;
-  firstPurchaseDate: string | null;
-  daysHeld: number | null;
-  pctOfPortfolio: number | null;
-};
-
-type ByHoldingRow = {
-  key: string;
-  symbol: string | null;
-  name: string;
-  assetType: AssetType;
-  totalQty: number;
-  avgCostDisplay: number | null;
-  costBasisDisplay: number;
-  marketValueDisplay: number;
-  unrealizedGainDisplay: number;
-  unrealizedGainPct: number | null;
-  realizedGainDisplay: number;
-  dividendsDisplay: number;
-  totalReturnDisplay: number;
-  totalReturnPct: number | null;
-  pctOfPortfolio: number | null;
-  accountCount: number;
-  image: string | null;
-};
-
-type AggregatedStock = {
-  ticker: string;
-  name: string;
-  sector: string;
-  country: string;
-  effectiveWeight: number;
-  effectiveValueDisplay: number;
-  contributingEtfs: { symbol: string; weight: number }[];
-};
-
-type EtfDetail = {
-  symbol: string;
-  name: string;
-  account: string;
-  fullName: string;
-  totalHoldings: number;
-  valueCAD: number;
-  weightPct: number;
-};
-
-type OverviewData = {
-  holdings: EnrichedHolding[];
-  byHolding?: ByHoldingRow[];
-  // Currency the API used for FX conversion + summary totals.
-  // marketValueDisplay field on each holding is denominated in this — the
-  // legacy "CAD" suffix on the field name is misleading, the value
-  // tracks the user's display currency.
-  displayCurrency?: string;
-  undecryptedTxCount?: number;
-  summary: {
-    totalHoldings: number;
-    totalAccounts: number;
-    totalValueDisplay: number;
-    dayChangeDisplay: number;
-    dayChangePct: number;
-    hasQuantityData: boolean;
-    totalCostBasisDisplay: number;
-    totalUnrealizedGainDisplay: number;
-    totalUnrealizedGainPct: number;
-    totalRealizedGainDisplay: number;
-    totalDividendsDisplay: number;
-    totalReturnDisplay: number;
-    totalReturnPct: number;
-  };
-  byType: Record<AssetType, { count: number; value: number }>;
-  byAccount: Record<string, { count: number; value: number }>;
-  etfXray: {
-    etfCount: number;
-    etfTotalValueDisplay: number;
-    etfs: EtfDetail[];
-    regions: Record<string, number>;
-    sectors: Record<string, number>;
-    aggregatedStocks: AggregatedStock[];
-  };
-  topGainers: EnrichedHolding[];
-  topLosers: EnrichedHolding[];
-};
-
-type BenchmarkData = {
-  symbol: string;
-  name: string;
-  color: string;
-  returnPct: number;
-  series: { date: string; value: number }[];
-};
-
-type FilterType = "all" | AssetType;
-type EtfXrayTab = "stocks" | "regions" | "sectors" | "etfs";
-
-// ── Tooltip Components ──────────────────────────────────────────────
-function GlassTooltip({
-  active, payload, label, formatter,
-}: {
-  active?: boolean;
-  payload?: { name: string; value: number; color?: string; payload?: Record<string, unknown> }[];
-  label?: string;
-  formatter?: (value: number, name: string) => string;
-}) {
-  if (!active || !payload?.length) return null;
-  return (
-    <div className="rounded-lg border border-border/50 bg-card/95 backdrop-blur-sm px-3 py-2 shadow-lg">
-      {label && <p className="text-xs text-muted-foreground mb-1">{label}</p>}
-      {payload.map((entry, i) => (
-        <div key={i} className="flex items-center gap-2 text-sm">
-          {entry.color && <ColorDot color={entry.color} />}
-          <span className="text-muted-foreground">{entry.name}:</span>
-          <span className="font-semibold">
-            {formatter ? formatter(entry.value, entry.name) : entry.value}
-          </span>
-        </div>
-      ))}
-    </div>
-  );
-}
-
-function ExposurePieTooltip({
-  active, payload,
-}: {
-  active?: boolean;
-  payload?: { name: string; value: number; payload: { name: string; pct: number } }[];
-}) {
-  if (!active || !payload?.length) return null;
-  const entry = payload[0];
-  return (
-    <div className="rounded-lg border border-border/50 bg-card/95 backdrop-blur-sm px-3 py-2 shadow-lg">
-      <p className="text-xs font-semibold">{entry.payload.name}</p>
-      <p className="text-sm font-bold">{entry.payload.pct}%</p>
-    </div>
-  );
-}
-
-// ── Change Badge ────────────────────────────────────────────────────
-function ChangeBadge({ value, className = "" }: { value: number | null; className?: string }) {
-  if (value === null || value === undefined) return <span className="text-muted-foreground">--</span>;
-  const isPositive = value >= 0;
-  return (
-    <span className={`inline-flex items-center gap-0.5 font-mono text-sm font-medium ${isPositive ? "text-emerald-600 dark:text-emerald-400" : "text-rose-600 dark:text-rose-400"} ${className}`}>
-      {isPositive ? <ArrowUpRight className="h-3.5 w-3.5" /> : <ArrowDownRight className="h-3.5 w-3.5" />}
-      {Math.abs(value).toFixed(2)}%
-    </span>
-  );
-}
-
-// Day-change as a percentage badge with the dollar amount (display currency)
-// shown inline beside it. `amount` is this holding's contribution to the
-// portfolio day change (change-per-unit × qty, FX-converted server-side).
-function DayChange({
-  pct,
-  amount,
-  currency,
-}: { pct: number | null; amount: number | null; currency: string }) {
-  if (pct === null || pct === undefined) return <span className="text-muted-foreground">--</span>;
-  return (
-    <span className="inline-flex items-center justify-end gap-1.5 whitespace-nowrap">
-      <ChangeBadge value={pct} />
-      {amount !== null && amount !== undefined && (
-        <span className={`text-xs font-mono ${amount >= 0 ? "text-emerald-600 dark:text-emerald-400" : "text-rose-600 dark:text-rose-400"}`}>
-          {amount >= 0 ? "+" : ""}{formatCurrency(amount, currency)}
-        </span>
-      )}
-    </span>
-  );
-}
-
-// ── CSV Export ──────────────────────────────────────────────────────
-function downloadCSV(csv: string, filename: string) {
-  const blob = new Blob([csv], { type: "text/csv" });
-  const url = URL.createObjectURL(blob);
-  const a = document.createElement("a");
-  a.href = url;
-  a.download = filename;
-  a.click();
-  URL.revokeObjectURL(url);
-}
-
-function exportStocksToCSV(stocks: AggregatedStock[], etfTotalValueDisplay: number, displayCurrency: string) {
-  const header = ["#", "Stock", "Ticker", "Sector", "Country", "Weight %", `Value ${displayCurrency}`, "Contributing ETFs", "ETF Weights"];
-  const rows = stocks.map((s, i) => [
-    i + 1,
-    `"${s.name}"`,
-    s.ticker,
-    s.sector,
-    s.country,
-    s.effectiveWeight.toFixed(2),
-    s.effectiveValueDisplay.toFixed(2),
-    `"${s.contributingEtfs.map(e => e.symbol).join(", ")}"`,
-    `"${s.contributingEtfs.map(e => `${e.symbol}: ${e.weight}%`).join(", ")}"`,
-  ]);
-  const totalWeight = stocks.reduce((s, x) => s + x.effectiveWeight, 0);
-  const totalValue = stocks.reduce((s, x) => s + x.effectiveValueDisplay, 0);
-  rows.push(["", "", "", "", "TOTAL", totalWeight.toFixed(2), totalValue.toFixed(2), "", ""] as unknown as string[]);
-  rows.push(["", "", "", "", "ETF Portfolio Value", "", etfTotalValueDisplay.toFixed(2), "", ""] as unknown as string[]);
-
-  const csv = [header.join(","), ...rows.map(r => (r as (string | number)[]).join(","))].join("\n");
-  downloadCSV(csv, `etf-stock-exposure-${new Date().toISOString().slice(0, 10)}.csv`);
-}
-
-// `exportHoldingsToCSV` was removed 2026-05-01 (issue #25): the All
-// Holdings table now renders canonical-holding rows from `byHolding`,
-// not per-(account, holding) rows from `holdings`. The CSV button uses
-// `exportByHoldingToCSV` below; users who want a per-account dump can
-// use the Holdings-by-Account panel's per-row drill-down or the
-// /api/portfolio endpoint.
-function exportByHoldingToCSV(rows: ByHoldingRow[], displayCurrency: string) {
-  const header = ["#", "Holding", "Symbol", "Type", "Total Qty", `Avg Cost ${displayCurrency}`, `Cost Basis ${displayCurrency}`, `Mkt Value ${displayCurrency}`, "Unrealized G/L", "Unrealized %", "Realized G/L", "Dividends", "Total Return", "Total Return %", "Accounts", "Weight %"];
-  const out = rows.map((r, i) => [
-    i + 1,
-    `"${r.name}"`,
-    r.symbol ?? "",
-    r.assetType,
-    r.totalQty,
-    r.avgCostDisplay?.toFixed(4) ?? "",
-    r.costBasisDisplay.toFixed(2),
-    r.marketValueDisplay.toFixed(2),
-    r.unrealizedGainDisplay.toFixed(2),
-    r.unrealizedGainPct?.toFixed(2) ?? "",
-    r.realizedGainDisplay.toFixed(2),
-    r.dividendsDisplay.toFixed(2),
-    r.totalReturnDisplay.toFixed(2),
-    r.totalReturnPct?.toFixed(2) ?? "",
-    r.accountCount,
-    r.pctOfPortfolio?.toFixed(2) ?? "",
-  ]);
-  const totalMV = rows.reduce((s, r) => s + r.marketValueDisplay, 0);
-  const totalUnreal = rows.reduce((s, r) => s + r.unrealizedGainDisplay, 0);
-  out.push(["", "TOTAL", "", "", "", "", "", totalMV.toFixed(2), totalUnreal.toFixed(2), "", "", "", "", "", "", "100.00"] as unknown as string[]);
-  const csv = [header.join(","), ...out.map(r => (r as (string | number | null)[]).join(","))].join("\n");
-  downloadCSV(csv, `portfolio-by-holding-${new Date().toISOString().slice(0, 10)}.csv`);
-}
-
-// ── Skeleton ────────────────────────────────────────────────────────
-function PortfolioSkeleton() {
-  return (
-    <div className="space-y-6">
-      <div>
-        <div className="h-8 w-48 bg-muted animate-shimmer rounded-lg" />
-        <div className="h-4 w-72 bg-muted animate-shimmer rounded-lg mt-2" />
-      </div>
-      <div className="grid grid-cols-2 lg:grid-cols-4 gap-4">
-        {Array.from({ length: 4 }).map((_, i) => (
-          <Card key={i}>
-            <CardContent className="pt-5">
-              <div className="space-y-2">
-                <div className="h-3 w-20 bg-muted animate-shimmer rounded" />
-                <div className="h-8 w-28 bg-muted animate-shimmer rounded" />
-                <div className="h-3 w-16 bg-muted animate-shimmer rounded" />
-              </div>
-            </CardContent>
-          </Card>
-        ))}
-      </div>
-      <Card>
-        <CardHeader><div className="h-5 w-36 bg-muted animate-shimmer rounded" /></CardHeader>
-        <CardContent className="space-y-3">
-          {Array.from({ length: 8 }).map((_, i) => (
-            <div key={i} className="flex items-center gap-4 py-2">
-              <div className="h-8 w-8 rounded-full bg-muted animate-shimmer" />
-              <div className="h-4 w-24 bg-muted animate-shimmer rounded" />
-              <div className="h-4 w-16 bg-muted animate-shimmer rounded ml-auto" />
-              <div className="h-4 w-16 bg-muted animate-shimmer rounded" />
-            </div>
-          ))}
-        </CardContent>
-      </Card>
-    </div>
-  );
-}
+import {
+  ASSET_TYPE_CONFIG, REGION_COLORS, SECTOR_COLORS,
+  clientCanonicalKey,
+  type EnrichedHolding, type EtfXrayTab, type FilterType,
+} from "./_types";
+import { usePortfolioOverview, useBenchmarks } from "./_hooks/use-portfolio";
+import { ChangeBadge, DayChange, PortfolioSkeleton } from "./_components/portfolio-ui";
+import { HoldingsTable } from "./_components/holdings-table";
+import { EtfXrayCard } from "./_components/etf-xray-card";
+import { AllocationCharts } from "./_components/allocation-charts";
+import { BenchmarkChart } from "./_components/benchmark-chart";
+import { HoldingsByAccount } from "./_components/holdings-by-account";
 
 // ── Main Page ───────────────────────────────────────────────────────
 export default function PortfolioPage() {
   const devMode = useDevMode();
   const { displayCurrency } = useDisplayCurrency();
-  const [data, setData] = useState<OverviewData | null>(null);
-  const [loading, setLoading] = useState(true);
+  const { data, loading, reload } = usePortfolioOverview(displayCurrency);
   const [filter, setFilter] = useState<FilterType>("all");
   // Edit/delete dialog for individual portfolio holdings. Null = closed.
   const [editingHolding, setEditingHolding] = useState<EnrichedHolding | null>(null);
@@ -389,9 +43,8 @@ export default function PortfolioPage() {
   // shared <HoldingEditForm> with the dropdown defaulted to that account.
   // {} = open with no default account (header-level "Add holding" button).
   const [creatingHolding, setCreatingHolding] = useState<{ accountId?: number } | null>(null);
-  const [benchmarks, setBenchmarks] = useState<BenchmarkData[]>([]);
   const [benchmarkPeriod, setBenchmarkPeriod] = useState("1y");
-  const [benchmarkLoading, setBenchmarkLoading] = useState(false);
+  const { benchmarks, benchmarkLoading } = useBenchmarks(benchmarkPeriod, !!devMode);
   const [expandedAccounts, setExpandedAccounts] = useState<Set<string>>(new Set());
   const [sortField, setSortField] = useState<"name" | "marketValueDisplay" | "unrealizedGainPct">("marketValueDisplay");
   const [sortDir, setSortDir] = useState<"asc" | "desc">("desc");
@@ -413,25 +66,6 @@ export default function PortfolioPage() {
   // safe). Default on — empty rows are noise on a value-weighted view.
   // Toggle lives in the holdings table filter bar.
   const [hideEmpty, setHideEmpty] = useState(true);
-
-  // Fetch portfolio overview — re-runs when display currency changes so
-  // totals + currency-as-holding prices reflect the user's choice.
-  useEffect(() => {
-    fetch(`/api/portfolio/overview?currency=${encodeURIComponent(displayCurrency)}`)
-      .then(r => r.json())
-      .then(d => { setData(d); setLoading(false); })
-      .catch(() => setLoading(false));
-  }, [displayCurrency]);
-
-  // Fetch benchmarks (dev mode only)
-  useEffect(() => {
-    if (!devMode) return;
-    setBenchmarkLoading(true);
-    fetch(`/api/portfolio/benchmarks?period=${benchmarkPeriod}`)
-      .then(r => r.json())
-      .then(d => { setBenchmarks(d.benchmarks ?? []); setBenchmarkLoading(false); })
-      .catch(() => setBenchmarkLoading(false));
-  }, [benchmarkPeriod, devMode]);
 
   // Issue #25 restructure: top-level All Holdings rows are canonical-
   // holding rollups (one row per ticker / cash sleeve / currency code),
@@ -505,11 +139,6 @@ export default function PortfolioPage() {
     });
   };
 
-  const SortIcon = ({ field }: { field: typeof sortField }) => {
-    if (sortField !== field) return null;
-    return sortDir === "asc" ? <ChevronUp className="h-3 w-3 inline ml-0.5" /> : <ChevronDown className="h-3 w-3 inline ml-0.5" />;
-  };
-
   if (loading) return <PortfolioSkeleton />;
   if (!data) return <div className="text-center py-12 text-muted-foreground">Unable to load portfolio data.</div>;
 
@@ -539,9 +168,6 @@ export default function PortfolioPage() {
   }
 
   const { summary, byType, etfXray, topGainers, topLosers } = data;
-
-  // Build benchmark chart data
-  const benchmarkChartData = buildBenchmarkChartData(benchmarks);
 
   // Allocation data — value-weighted in display currency. `value` powers
   // both slice size and legend amount; we filter zero-value buckets when
@@ -838,845 +464,54 @@ export default function PortfolioPage() {
       )}
 
       {/* ── Holdings Table ────────────────────────────────────── */}
-      <Card>
-        <CardHeader className="pb-2">
-          <div className="flex flex-col sm:flex-row sm:items-center justify-between gap-3">
-            <div>
-              <CardTitle className="text-base">All Holdings</CardTitle>
-              <p className="text-xs text-muted-foreground mt-0.5">
-                Live prices from Yahoo Finance &amp; CoinGecko
-              </p>
-            </div>
-            <div className="flex items-center gap-2">
-              <div className="flex gap-1 flex-wrap">
-                {(["all", "etf", "stock", "crypto", "cash"] as const).map(t => (
-                  <Button
-                    key={t}
-                    variant={filter === t ? "default" : "outline"}
-                    size="sm"
-                    className="text-xs h-7 px-2.5"
-                    onClick={() => setFilter(t)}
-                  >
-                    {t === "all" ? "All" : ASSET_TYPE_CONFIG[t]?.label ?? t}
-                    <Badge variant="secondary" className="ml-1 text-[10px] h-4 px-1 bg-transparent">
-                      {t === "all" ? summary.totalHoldings : byType[t]?.count ?? 0}
-                    </Badge>
-                  </Button>
-                ))}
-              </div>
-              <Button
-                variant={hideEmpty ? "default" : "outline"}
-                size="sm"
-                className="text-xs gap-1.5 h-7"
-                onClick={() => setHideEmpty(!hideEmpty)}
-                title="Hide holdings with no current position (quantity = 0, e.g. fully-sold)"
-              >
-                {hideEmpty ? "Hiding empty" : "Showing all"}
-              </Button>
-              <Button
-                variant="outline"
-                size="sm"
-                className="text-xs gap-1.5 h-7"
-                onClick={() => exportByHoldingToCSV(data.byHolding ?? [], data.displayCurrency ?? displayCurrency)}
-              >
-                <Download className="h-3 w-3" />
-                CSV
-              </Button>
-            </div>
-          </div>
-        </CardHeader>
-        <CardContent>
-          <div className="overflow-x-auto">
-            <Table>
-              <TableHeader>
-                <TableRow>
-                  <TableHead className="cursor-pointer select-none w-8" />
-                  <TableHead className="cursor-pointer select-none" onClick={() => handleSort("name")}>
-                    Holding <SortIcon field="name" />
-                  </TableHead>
-                  <TableHead className="text-right">Total Qty</TableHead>
-                  <TableHead className="text-right">Avg Cost</TableHead>
-                  <TableHead className="text-right cursor-pointer select-none" onClick={() => handleSort("marketValueDisplay")}>
-                    Mkt Value <SortIcon field="marketValueDisplay" />
-                  </TableHead>
-                  <TableHead className="text-right cursor-pointer select-none" onClick={() => handleSort("unrealizedGainPct")}>
-                    Unrealized G/L <SortIcon field="unrealizedGainPct" />
-                  </TableHead>
-                  <TableHead className="text-right">Realized G/L</TableHead>
-                  <TableHead className="text-right">Accounts</TableHead>
-                </TableRow>
-              </TableHeader>
-              <TableBody>
-                {filteredHoldings.map(r => {
-                  const typeConf = ASSET_TYPE_CONFIG[r.assetType];
-                  const isExpanded = expandedRows.has(r.key);
-                  const reportCcy = data.displayCurrency ?? displayCurrency;
-                  const memberHoldings = holdingsByCanonicalKey.get(r.key) ?? [];
-                  // Aggregate-level first-purchase / days-held are derived
-                  // from the per-account members so they reflect the
-                  // earliest buy across every account that holds this
-                  // canonical position.
-                  const earliestPurchase = memberHoldings.reduce<string | null>((acc, h) => {
-                    if (!h.firstPurchaseDate) return acc;
-                    if (!acc || h.firstPurchaseDate < acc) return h.firstPurchaseDate;
-                    return acc;
-                  }, null);
-                  const today = new Date();
-                  const daysHeld = earliestPurchase
-                    ? Math.floor((today.getTime() - new Date(earliestPurchase).getTime()) / 86400000)
-                    : null;
-                  return (
-                    <>
-                      <TableRow
-                        key={r.key}
-                        className="hover:bg-muted/30 transition-colors cursor-pointer"
-                        onClick={() => toggleRow(r.key)}
-                      >
-                        <TableCell className="text-muted-foreground">
-                          {isExpanded ? <ChevronUp className="h-3 w-3" /> : <ChevronDown className="h-3 w-3" />}
-                        </TableCell>
-                        <TableCell>
-                          <div className="flex items-center gap-2">
-                            {r.image && <img src={r.image} alt="" className="h-6 w-6 rounded-full" />}
-                            <div>
-                              <span className="font-medium text-sm">{r.name}</span>
-                              <div className="flex items-center gap-1 mt-0.5">
-                                {r.symbol && <Badge variant="secondary" className="font-mono text-[10px] h-4 px-1">{r.symbol}</Badge>}
-                                <Badge
-                                  variant="outline"
-                                  className="text-[10px] h-4 px-1"
-                                  ref={(el) => {
-                                    if (el && typeConf?.color) {
-                                      el.style.borderColor = typeConf.color;
-                                      el.style.color = typeConf.color;
-                                    }
-                                  }}
-                                >
-                                  {typeConf?.label ?? r.assetType}
-                                </Badge>
-                                {r.totalQty < 0 && (
-                                  <Badge
-                                    variant="outline"
-                                    className="text-[10px] h-4 px-1 border-rose-500 text-rose-600 dark:border-rose-400 dark:text-rose-400"
-                                    title="Net-short position — sales exceeded buys. Lots are tracked via holding_lots.side='short'; close by buying to cover."
-                                  >
-                                    Short
-                                  </Badge>
-                                )}
-                              </div>
-                            </div>
-                          </div>
-                        </TableCell>
-                        <TableCell className={`text-right font-mono text-sm ${r.totalQty < 0 ? "text-rose-600 dark:text-rose-400" : ""}`}>
-                          {r.totalQty !== 0
-                            ? r.totalQty.toLocaleString("en-CA", { minimumFractionDigits: 0, maximumFractionDigits: r.totalQty % 1 === 0 ? 0 : 4 })
-                            : <span className="text-muted-foreground text-xs">--</span>}
-                        </TableCell>
-                        <TableCell className="text-right font-mono text-sm">
-                          {r.avgCostDisplay != null ? formatCurrency(r.avgCostDisplay, reportCcy) : <span className="text-muted-foreground text-xs">--</span>}
-                        </TableCell>
-                        <TableCell className="text-right font-mono text-sm font-medium">
-                          {r.marketValueDisplay !== 0 ? formatCurrency(r.marketValueDisplay, reportCcy) : <span className="text-muted-foreground text-xs">--</span>}
-                        </TableCell>
-                        <TableCell className="text-right">
-                          <div>
-                            <p className={`text-sm font-mono font-medium ${r.unrealizedGainDisplay >= 0 ? "text-emerald-600 dark:text-emerald-400" : "text-rose-600 dark:text-rose-400"}`}>
-                              {r.unrealizedGainDisplay >= 0 ? "+" : ""}{formatCurrency(r.unrealizedGainDisplay, reportCcy)}
-                            </p>
-                            {r.unrealizedGainPct != null && (
-                              <p className={`text-xs font-mono ${r.unrealizedGainPct >= 0 ? "text-emerald-600 dark:text-emerald-400" : "text-rose-600 dark:text-rose-400"}`}>
-                                {r.unrealizedGainPct >= 0 ? "+" : ""}{r.unrealizedGainPct.toFixed(2)}%
-                              </p>
-                            )}
-                          </div>
-                        </TableCell>
-                        <TableCell className="text-right font-mono text-sm">
-                          {r.realizedGainDisplay !== 0 ? (
-                            <span className={`${r.realizedGainDisplay >= 0 ? "text-emerald-600 dark:text-emerald-400" : "text-rose-600 dark:text-rose-400"}`}>
-                              {r.realizedGainDisplay >= 0 ? "+" : ""}{formatCurrency(r.realizedGainDisplay, reportCcy)}
-                            </span>
-                          ) : <span className="text-muted-foreground text-xs">--</span>}
-                        </TableCell>
-                        <TableCell className="text-right">
-                          <Badge variant="outline" className="text-[10px]">
-                            {r.accountCount}
-                          </Badge>
-                        </TableCell>
-                      </TableRow>
-                      {isExpanded && (
-                        <TableRow key={`${r.key}-detail`} className="bg-muted/10 border-0">
-                          <TableCell />
-                          <TableCell colSpan={7} className="py-3">
-                            {/* Aggregate-level info grid (shared across the
-                                accounts inside this canonical position). */}
-                            <div className="grid grid-cols-2 sm:grid-cols-3 md:grid-cols-6 gap-3 text-xs">
-                              <div>
-                                <p className="text-muted-foreground">First Purchase</p>
-                                <p className="font-medium">{earliestPurchase ?? "--"}</p>
-                              </div>
-                              <div>
-                                <p className="text-muted-foreground">Days Held</p>
-                                <p className="font-medium">{daysHeld != null ? `${daysHeld.toLocaleString()} days` : "--"}</p>
-                              </div>
-                              <div>
-                                <p className="text-muted-foreground">% of Portfolio</p>
-                                <p className="font-medium">{r.pctOfPortfolio != null ? `${r.pctOfPortfolio.toFixed(2)}%` : "--"}</p>
-                              </div>
-                              <div>
-                                <p className="text-muted-foreground">Cost Basis</p>
-                                <p className="font-medium font-mono">{r.costBasisDisplay !== 0 ? formatCurrency(r.costBasisDisplay, reportCcy) : "--"}</p>
-                              </div>
-                              <div>
-                                <p className="text-muted-foreground">Dividends</p>
-                                <p className="font-medium font-mono text-emerald-600 dark:text-emerald-400">
-                                  {r.dividendsDisplay > 0 ? `+${formatCurrency(r.dividendsDisplay, reportCcy)}` : "--"}
-                                </p>
-                              </div>
-                              <div>
-                                <p className="text-muted-foreground">Total Return</p>
-                                <p className={`font-medium font-mono ${r.totalReturnDisplay >= 0 ? "text-emerald-600 dark:text-emerald-400" : "text-rose-600 dark:text-rose-400"}`}>
-                                  {r.totalReturnDisplay !== 0 ? `${r.totalReturnDisplay >= 0 ? "+" : ""}${formatCurrency(r.totalReturnDisplay, reportCcy)}` : "--"}
-                                  {r.totalReturnPct != null && (
-                                    <span className="ml-1 text-[10px]">({r.totalReturnPct >= 0 ? "+" : ""}{r.totalReturnPct.toFixed(1)}%)</span>
-                                  )}
-                                </p>
-                              </div>
-                            </div>
-
-                            {/* Per-account breakdown (issue #25 spec). One
-                                row per (canonical-holding, account) pair
-                                with a "View transactions" drill-down. Always
-                                rendered — for single-account holdings it
-                                doubles as the only edit affordance, since
-                                the account-name cell opens the editor. */}
-                            {memberHoldings.length > 0 && (
-                              <div className="mt-3 rounded-md border border-border overflow-hidden">
-                                <Table>
-                                  <TableHeader className="bg-muted/40">
-                                    <TableRow className="hover:bg-transparent border-border">
-                                      <TableHead className="text-[10px] uppercase tracking-wider text-muted-foreground">Account</TableHead>
-                                      <TableHead className="text-right text-[10px] uppercase tracking-wider text-muted-foreground">Qty</TableHead>
-                                      <TableHead className="text-right text-[10px] uppercase tracking-wider text-muted-foreground">Avg Cost</TableHead>
-                                      <TableHead className="text-right text-[10px] uppercase tracking-wider text-muted-foreground">Mkt Value</TableHead>
-                                      <TableHead className="text-right text-[10px] uppercase tracking-wider text-muted-foreground">Unrealized G/L</TableHead>
-                                      <TableHead className="text-right text-[10px] uppercase tracking-wider text-muted-foreground">Realized G/L</TableHead>
-                                      <TableHead className="text-right" />
-                                    </TableRow>
-                                  </TableHeader>
-                                  <TableBody>
-                                    {memberHoldings.map(h => {
-                                      const hasMetrics = h.quantity !== null && h.quantity !== 0;
-                                      const nativeCcy = h.quoteCurrency ?? h.currency;
-                                      return (
-                                        <TableRow key={h.id} className="hover:bg-muted/30 border-border/60">
-                                          <TableCell className="text-xs">
-                                            <button
-                                              type="button"
-                                              className="text-left hover:underline"
-                                              onClick={(e) => { e.stopPropagation(); setEditingHolding(h); }}
-                                              title="Edit this per-account holding row"
-                                            >
-                                              {h.accountName}
-                                            </button>
-                                          </TableCell>
-                                          <TableCell className={`text-right font-mono text-xs ${hasMetrics && h.quantity != null && h.quantity < 0 ? "text-rose-600 dark:text-rose-400" : ""}`}>
-                                            {hasMetrics && h.quantity != null
-                                              ? h.quantity.toLocaleString("en-CA", { minimumFractionDigits: 0, maximumFractionDigits: h.quantity % 1 === 0 ? 0 : 4 })
-                                              : <span className="text-muted-foreground">--</span>}
-                                            {hasMetrics && h.quantity != null && h.quantity < 0 && (
-                                              <span className="ml-1 text-[9px] uppercase tracking-wider text-rose-500" title="Short position">short</span>
-                                            )}
-                                          </TableCell>
-                                          <TableCell className="text-right font-mono text-xs">
-                                            {hasMetrics && h.avgCostPerShare != null
-                                              ? formatCurrency(h.avgCostPerShare, nativeCcy)
-                                              : <span className="text-muted-foreground">--</span>}
-                                          </TableCell>
-                                          <TableCell className="text-right font-mono text-xs font-medium">
-                                            {hasMetrics && h.marketValueDisplay != null
-                                              ? formatCurrency(h.marketValueDisplay, reportCcy)
-                                              : <span className="text-muted-foreground">--</span>}
-                                          </TableCell>
-                                          <TableCell className="text-right text-xs">
-                                            {hasMetrics && h.unrealizedGain != null ? (
-                                              <span className={`font-mono ${h.unrealizedGain >= 0 ? "text-emerald-600 dark:text-emerald-400" : "text-rose-600 dark:text-rose-400"}`}>
-                                                {h.unrealizedGain >= 0 ? "+" : ""}{formatCurrency(h.unrealizedGain, nativeCcy)}
-                                              </span>
-                                            ) : <span className="text-muted-foreground">--</span>}
-                                          </TableCell>
-                                          <TableCell className="text-right text-xs">
-                                            {h.realizedGain != null && h.realizedGain !== 0 ? (
-                                              <span className={`font-mono ${h.realizedGain >= 0 ? "text-emerald-600 dark:text-emerald-400" : "text-rose-600 dark:text-rose-400"}`}>
-                                                {h.realizedGain >= 0 ? "+" : ""}{formatCurrency(h.realizedGain, nativeCcy)}
-                                              </span>
-                                            ) : <span className="text-muted-foreground">--</span>}
-                                          </TableCell>
-                                          <TableCell className="text-right">
-                                            <Link
-                                              href={`/transactions?portfolioHolding=${encodeURIComponent(h.name)}${h.accountId ? `&accountId=${h.accountId}` : ""}`}
-                                              onClick={(e) => e.stopPropagation()}
-                                              className="text-[11px] text-primary hover:underline whitespace-nowrap"
-                                              title="View transactions for this holding in this account"
-                                            >
-                                              View txns →
-                                            </Link>
-                                          </TableCell>
-                                        </TableRow>
-                                      );
-                                    })}
-                                  </TableBody>
-                                </Table>
-                              </div>
-                            )}
-
-                            <div className="mt-3 pt-3 border-t border-border/50 flex justify-end gap-3">
-                              {/* "View transactions" — same canonical
-                                  holding across every account it lives in.
-                                  Mirrors the existing
-                                  /transactions?portfolioHolding=<name>
-                                  contract; the per-account "View txns"
-                                  links above scope to a single account.
-                                  Per-account Edit lives on the account-name
-                                  cell in the breakdown table above. */}
-                              <Link
-                                href={`/transactions?portfolioHolding=${encodeURIComponent(r.name)}`}
-                                onClick={(e) => e.stopPropagation()}
-                                className="inline-flex items-center gap-1.5 text-xs text-primary hover:underline font-medium"
-                              >
-                                View transactions →
-                              </Link>
-                            </div>
-                          </TableCell>
-                        </TableRow>
-                      )}
-                    </>
-                  );
-                })}
-                {filteredHoldings.length === 0 && (
-                  <TableRow>
-                    <TableCell colSpan={8} className="text-center text-muted-foreground py-8">
-                      No {filter === "all" ? "" : ASSET_TYPE_CONFIG[filter]?.label} holdings found.
-                      {hideEmpty && data.holdings.length > 0 && (
-                        <span className="block mt-1 text-xs">
-                          Showing only positions with quantity &gt; 0.
-                          <button
-                            onClick={() => setHideEmpty(false)}
-                            className="ml-1 underline hover:text-foreground"
-                          >
-                            Show all
-                          </button>
-                        </span>
-                      )}
-                    </TableCell>
-                  </TableRow>
-                )}
-              </TableBody>
-            </Table>
-          </div>
-        </CardContent>
-      </Card>
+      <HoldingsTable
+        data={data}
+        displayCurrency={displayCurrency}
+        filteredHoldings={filteredHoldings}
+        holdingsByCanonicalKey={holdingsByCanonicalKey}
+        filter={filter}
+        setFilter={setFilter}
+        hideEmpty={hideEmpty}
+        setHideEmpty={setHideEmpty}
+        sortField={sortField}
+        sortDir={sortDir}
+        handleSort={handleSort}
+        expandedRows={expandedRows}
+        toggleRow={toggleRow}
+        setEditingHolding={setEditingHolding}
+      />
 
       {/* ── ETF X-Ray (Combined) — dev only ──────────────────── */}
       {devMode && hasEtfData && (
-        <Card>
-          <CardHeader className="pb-2">
-            <div className="flex flex-col sm:flex-row sm:items-center justify-between gap-3">
-              <div>
-                <div className="flex items-center gap-2">
-                  <Search className="h-4 w-4 text-indigo-500" />
-                  <CardTitle className="text-base">ETF X-Ray</CardTitle>
-                </div>
-                <p className="text-xs text-muted-foreground mt-0.5">
-                  Effective exposure across {etfXray.etfCount} ETF{etfXray.etfCount !== 1 ? "s" : ""} in your portfolio
-                </p>
-              </div>
-              <div className="flex gap-1">
-                {([
-                  { key: "stocks" as const, label: "Stocks", icon: BarChart3 },
-                  { key: "regions" as const, label: "Regions", icon: Globe2 },
-                  { key: "sectors" as const, label: "Sectors", icon: Building2 },
-                  { key: "etfs" as const, label: "Per ETF", icon: Layers },
-                ]).map(tab => (
-                  <Button
-                    key={tab.key}
-                    variant={etfXrayTab === tab.key ? "default" : "outline"}
-                    size="sm"
-                    className="text-xs h-7 px-2.5"
-                    onClick={() => setEtfXrayTab(tab.key)}
-                  >
-                    <tab.icon className="h-3 w-3 mr-1" />
-                    {tab.label}
-                  </Button>
-                ))}
-              </div>
-            </div>
-          </CardHeader>
-          <CardContent>
-            {/* ── Stocks Tab: Aggregated look-through ── */}
-            {etfXrayTab === "stocks" && (
-              <div className="space-y-4">
-                <p className="text-xs text-muted-foreground">
-                  Your effective stock exposure across all ETFs, weighted by each ETF&apos;s portfolio allocation.
-                </p>
-                <div className="overflow-x-auto">
-                  <Table>
-                    <TableHeader>
-                      <TableRow>
-                        <TableHead className="w-8">#</TableHead>
-                        <TableHead>Stock</TableHead>
-                        <TableHead>Ticker</TableHead>
-                        <TableHead>Sector</TableHead>
-                        <TableHead>Country</TableHead>
-                        <TableHead className="text-right">Weight</TableHead>
-                        <TableHead className="text-right">Value (CAD)</TableHead>
-                        <TableHead className="w-28">Exposure</TableHead>
-                        <TableHead>Via ETFs</TableHead>
-                      </TableRow>
-                    </TableHeader>
-                    <TableBody>
-                      {etfXray.aggregatedStocks
-                        .slice((stocksPage - 1) * STOCKS_PER_PAGE, stocksPage * STOCKS_PER_PAGE)
-                        .map((s, i) => {
-                          const globalIdx = (stocksPage - 1) * STOCKS_PER_PAGE + i;
-                          return (
-                            <TableRow key={s.ticker} className={`hover:bg-muted/30 transition-colors ${s.ticker === "OTHER" ? "bg-muted/20 border-t" : ""}`}>
-                              <TableCell className="text-xs text-muted-foreground font-mono">{globalIdx + 1}</TableCell>
-                              <TableCell className={`text-sm ${s.ticker === "OTHER" ? "italic text-muted-foreground" : "font-medium"}`}>{s.name}</TableCell>
-                              <TableCell>
-                                {s.ticker !== "OTHER" && <Badge variant="secondary" className="font-mono text-xs">{s.ticker}</Badge>}
-                              </TableCell>
-                              <TableCell>
-                                <Badge
-                                  variant="outline"
-                                  className="text-[10px]"
-                                  ref={(el: HTMLElement | null) => {
-                                    if (el) {
-                                      const c = SECTOR_COLORS[s.sector] ?? "#64748b";
-                                      el.style.borderColor = c;
-                                      el.style.color = c;
-                                    }
-                                  }}
-                                >
-                                  {s.sector}
-                                </Badge>
-                              </TableCell>
-                              <TableCell className="text-xs text-muted-foreground">{s.country}</TableCell>
-                              <TableCell className="text-right">
-                                <span className="text-sm font-mono font-semibold">{s.effectiveWeight.toFixed(1)}%</span>
-                              </TableCell>
-                              <TableCell className="text-right">
-                                <span className="text-sm font-mono text-muted-foreground">{formatCurrency(s.effectiveValueDisplay, "CAD")}</span>
-                              </TableCell>
-                              <TableCell>
-                                <div className="w-full h-1.5 rounded-full bg-muted overflow-hidden">
-                                  <div
-                                    className="h-full rounded-full bg-indigo-500"
-                                    ref={(el) => {
-                                      if (el) el.style.width = `${Math.min(s.effectiveWeight * 10, 100)}%`;
-                                    }}
-                                  />
-                                </div>
-                              </TableCell>
-                              <TableCell>
-                                <div className="flex gap-1 flex-wrap">
-                                  {s.contributingEtfs.map((e, ei) => (
-                                    <span key={`${e.symbol}-${ei}`} className="text-[10px] font-mono text-muted-foreground bg-muted px-1 py-0.5 rounded">
-                                      {e.symbol} {e.weight}%
-                                    </span>
-                                  ))}
-                                </div>
-                              </TableCell>
-                            </TableRow>
-                          );
-                        })}
-                    </TableBody>
-                  </Table>
-                </div>
-                {etfXray.aggregatedStocks.length > 0 && (() => {
-                  const totalPages = Math.ceil(etfXray.aggregatedStocks.length / STOCKS_PER_PAGE);
-                  const totalWeight = etfXray.aggregatedStocks.reduce((s, x) => s + x.effectiveWeight, 0);
-                  return (
-                    <div className="flex flex-col sm:flex-row items-start sm:items-center justify-between gap-2">
-                      <p className="text-xs text-muted-foreground">
-                        {etfXray.aggregatedStocks.length} stocks · Total weight: {totalWeight.toFixed(1)}%
-                      </p>
-                      <div className="flex items-center gap-2">
-                        {totalPages > 1 && (
-                          <div className="flex items-center gap-1">
-                            <Button
-                              variant="outline"
-                              size="sm"
-                              className="h-7 w-7 p-0"
-                              disabled={stocksPage <= 1}
-                              onClick={() => setStocksPage(p => p - 1)}
-                            >
-                              <ChevronLeft className="h-3 w-3" />
-                            </Button>
-                            <span className="text-xs text-muted-foreground px-1">
-                              {stocksPage} / {totalPages}
-                            </span>
-                            <Button
-                              variant="outline"
-                              size="sm"
-                              className="h-7 w-7 p-0"
-                              disabled={stocksPage >= totalPages}
-                              onClick={() => setStocksPage(p => p + 1)}
-                            >
-                              <ChevronRight className="h-3 w-3" />
-                            </Button>
-                          </div>
-                        )}
-                        <Button
-                          variant="outline"
-                          size="sm"
-                          className="text-xs gap-1.5"
-                          onClick={() => exportStocksToCSV(etfXray.aggregatedStocks, etfXray.etfTotalValueDisplay, displayCurrency)}
-                        >
-                          <Download className="h-3 w-3" />
-                          Export CSV
-                        </Button>
-                      </div>
-                    </div>
-                  );
-                })()}
-              </div>
-            )}
-
-            {/* ── Regions Tab ── */}
-            {etfXrayTab === "regions" && (
-              <div className="space-y-4">
-                {regionData.length > 0 ? (
-                  <div className="flex flex-col sm:flex-row items-start sm:items-center gap-6">
-                    <div className="w-48 h-48 shrink-0">
-                      <ResponsiveContainer width="100%" height="100%">
-                        <PieChart>
-                          <Pie
-                            data={regionData}
-                            dataKey="pct"
-                            nameKey="name"
-                            cx="50%"
-                            cy="50%"
-                            innerRadius={45}
-                            outerRadius={85}
-                            strokeWidth={2}
-                            stroke="var(--color-card)"
-                          >
-                            {regionData.map((d, i) => (
-                              <Cell key={i} fill={d.color} />
-                            ))}
-                          </Pie>
-                          <Tooltip content={<ExposurePieTooltip />} />
-                        </PieChart>
-                      </ResponsiveContainer>
-                    </div>
-                    <div className="flex-1 space-y-2 min-w-0">
-                      {regionData.map(d => (
-                        <div key={d.name} className="flex items-center gap-2">
-                          <ColorDot color={d.color} className="h-3 w-3" />
-                          <span className="text-sm text-muted-foreground flex-1">{d.name}</span>
-                          <div className="flex items-center gap-2">
-                            <div className="w-24 h-2 rounded-full bg-muted overflow-hidden">
-                              <CspSafeColorBar percent={d.pct} color={d.color} />
-                            </div>
-                            <span className="text-sm font-mono font-semibold w-12 text-right">{d.pct}%</span>
-                          </div>
-                        </div>
-                      ))}
-                    </div>
-                  </div>
-                ) : (
-                  <p className="text-sm text-muted-foreground text-center py-8">No region data available.</p>
-                )}
-              </div>
-            )}
-
-            {/* ── Sectors Tab ── */}
-            {etfXrayTab === "sectors" && (
-              <div className="space-y-4">
-                {sectorData.length > 0 ? (
-                  <div className="flex flex-col sm:flex-row items-start sm:items-center gap-6">
-                    <div className="w-48 h-48 shrink-0">
-                      <ResponsiveContainer width="100%" height="100%">
-                        <PieChart>
-                          <Pie
-                            data={sectorData}
-                            dataKey="pct"
-                            nameKey="name"
-                            cx="50%"
-                            cy="50%"
-                            innerRadius={45}
-                            outerRadius={85}
-                            strokeWidth={2}
-                            stroke="var(--color-card)"
-                          >
-                            {sectorData.map((d, i) => (
-                              <Cell key={i} fill={d.color} />
-                            ))}
-                          </Pie>
-                          <Tooltip content={<ExposurePieTooltip />} />
-                        </PieChart>
-                      </ResponsiveContainer>
-                    </div>
-                    <div className="flex-1 space-y-2 min-w-0">
-                      {sectorData.map(d => (
-                        <div key={d.name} className="flex items-center gap-2">
-                          <ColorDot color={d.color} className="h-3 w-3" />
-                          <span className="text-sm text-muted-foreground flex-1">{d.name}</span>
-                          <div className="flex items-center gap-2">
-                            <div className="w-24 h-2 rounded-full bg-muted overflow-hidden">
-                              <CspSafeColorBar percent={d.pct} color={d.color} />
-                            </div>
-                            <span className="text-sm font-mono font-semibold w-12 text-right">{d.pct}%</span>
-                          </div>
-                        </div>
-                      ))}
-                    </div>
-                  </div>
-                ) : (
-                  <p className="text-sm text-muted-foreground text-center py-8">No sector data available.</p>
-                )}
-              </div>
-            )}
-
-            {/* ── Per ETF Tab ── */}
-            {etfXrayTab === "etfs" && (
-              <div className="space-y-3">
-                <p className="text-xs text-muted-foreground">
-                  Your ETF holdings and their weight in the ETF portfolio.
-                </p>
-                <Table>
-                  <TableHeader>
-                    <TableRow>
-                      <TableHead>ETF</TableHead>
-                      <TableHead>Fund Name</TableHead>
-                      <TableHead>Account</TableHead>
-                      <TableHead className="text-right">Holdings</TableHead>
-                      <TableHead className="text-right">Portfolio Weight</TableHead>
-                      <TableHead className="w-24"></TableHead>
-                    </TableRow>
-                  </TableHeader>
-                  <TableBody>
-                    {etfXray.etfs.map(etf => (
-                      <TableRow key={`${etf.symbol}-${etf.account}`} className="hover:bg-muted/30 transition-colors">
-                        <TableCell>
-                          <Badge variant="secondary" className="font-mono text-xs">{etf.symbol}</Badge>
-                        </TableCell>
-                        <TableCell>
-                          <span className="text-sm font-medium">{etf.fullName}</span>
-                        </TableCell>
-                        <TableCell className="text-xs text-muted-foreground">{etf.account}</TableCell>
-                        <TableCell className="text-right text-sm font-mono text-muted-foreground">
-                          {etf.totalHoldings.toLocaleString()}
-                        </TableCell>
-                        <TableCell className="text-right">
-                          <span className="text-sm font-mono font-semibold">{etf.weightPct}%</span>
-                        </TableCell>
-                        <TableCell>
-                          <div className="w-full h-1.5 rounded-full bg-muted overflow-hidden">
-                            <div
-                              className="h-full rounded-full bg-indigo-500"
-                              ref={(el) => {
-                                if (el) el.style.width = `${Math.min(etf.weightPct, 100)}%`;
-                              }}
-                            />
-                          </div>
-                        </TableCell>
-                      </TableRow>
-                    ))}
-                  </TableBody>
-                </Table>
-              </div>
-            )}
-          </CardContent>
-        </Card>
+        <EtfXrayCard
+          etfXray={etfXray}
+          etfXrayTab={etfXrayTab}
+          setEtfXrayTab={setEtfXrayTab}
+          stocksPage={stocksPage}
+          setStocksPage={setStocksPage}
+          stocksPerPage={STOCKS_PER_PAGE}
+          regionData={regionData}
+          sectorData={sectorData}
+          displayCurrency={displayCurrency}
+        />
       )}
 
       {/* ── Allocation Overview ────────────────────────────────── */}
-      <div className="grid grid-cols-1 lg:grid-cols-2 gap-4">
-        {/* By Asset Type */}
-        <Card>
-          <CardHeader className="pb-2">
-            <div className="flex items-center gap-2">
-              <PieChartIcon className="h-4 w-4 text-indigo-500" />
-              <CardTitle className="text-base">By Asset Type</CardTitle>
-            </div>
-          </CardHeader>
-          <CardContent>
-            <div className="flex items-center gap-6">
-              <div className="w-36 h-36 shrink-0">
-                <ResponsiveContainer width="100%" height="100%">
-                  <PieChart>
-                    <Pie
-                      data={allocationByType}
-                      dataKey="value"
-                      nameKey="name"
-                      cx="50%"
-                      cy="50%"
-                      innerRadius={36}
-                      outerRadius={64}
-                      strokeWidth={2}
-                      stroke="var(--color-card)"
-                    >
-                      {allocationByType.map((d, i) => (
-                        <Cell key={i} fill={d.color} />
-                      ))}
-                    </Pie>
-                    <Tooltip content={<ExposurePieTooltip />} />
-                  </PieChart>
-                </ResponsiveContainer>
-              </div>
-              <div className="flex-1 space-y-2 min-w-0">
-                {allocationByType.map(d => (
-                  <div key={d.name} className="flex items-center gap-2">
-                    <ColorDot color={d.color} className="h-2.5 w-2.5" />
-                    <span className="text-xs text-muted-foreground flex-1">{d.name}</span>
-                    <span className="text-xs font-medium tabular-nums">{formatCurrency(d.value, displayCurrency)} ({d.pct}%)</span>
-                  </div>
-                ))}
-              </div>
-            </div>
-          </CardContent>
-        </Card>
-
-        {/* By Account */}
-        <Card>
-          <CardHeader className="pb-2">
-            <div className="flex items-center gap-2">
-              <Wallet className="h-4 w-4 text-violet-500" />
-              <CardTitle className="text-base">By Account</CardTitle>
-            </div>
-          </CardHeader>
-          <CardContent>
-            <div className="flex items-center gap-6">
-              <div className="w-36 h-36 shrink-0">
-                <ResponsiveContainer width="100%" height="100%">
-                  <PieChart>
-                    <Pie
-                      data={allocationByAccount}
-                      dataKey="value"
-                      nameKey="name"
-                      cx="50%"
-                      cy="50%"
-                      innerRadius={36}
-                      outerRadius={64}
-                      strokeWidth={2}
-                      stroke="var(--color-card)"
-                    >
-                      {allocationByAccount.map((_, i) => (
-                        <Cell key={i} fill={PIE_COLORS[i % PIE_COLORS.length]} />
-                      ))}
-                    </Pie>
-                    <Tooltip content={<ExposurePieTooltip />} />
-                  </PieChart>
-                </ResponsiveContainer>
-              </div>
-              <div className="flex-1 space-y-1.5 min-w-0 max-h-36 overflow-y-auto">
-                {allocationByAccount.map((d, i) => (
-                  <div key={d.name} className="flex items-center gap-2">
-                    <ColorDot color={PIE_COLORS[i % PIE_COLORS.length]} className="h-2.5 w-2.5" />
-                    <span className="text-xs text-muted-foreground flex-1 truncate">{d.name}</span>
-                    <span className="text-xs font-medium tabular-nums">{formatCurrency(d.value, displayCurrency)} ({d.pct}%)</span>
-                  </div>
-                ))}
-              </div>
-            </div>
-          </CardContent>
-        </Card>
-      </div>
+      <AllocationCharts
+        allocationByType={allocationByType}
+        allocationByAccount={allocationByAccount}
+        displayCurrency={displayCurrency}
+      />
 
       {/* ── Performance vs Benchmarks — dev only ─────────────── */}
-      {devMode && <Card>
-        <CardHeader className="pb-2">
-          <div className="flex items-center justify-between">
-            <div>
-              <CardTitle className="text-base">Performance vs Benchmarks</CardTitle>
-              <p className="text-xs text-muted-foreground mt-0.5">
-                Compare major indices over time (% change)
-              </p>
-            </div>
-            <div className="flex gap-1">
-              {(["ytd", "1y", "3y", "5y"] as const).map(p => (
-                <Button
-                  key={p}
-                  variant={benchmarkPeriod === p ? "default" : "outline"}
-                  size="sm"
-                  className="text-xs h-7 px-2"
-                  onClick={() => setBenchmarkPeriod(p)}
-                >
-                  {p.toUpperCase()}
-                </Button>
-              ))}
-            </div>
-          </div>
-        </CardHeader>
-        <CardContent>
-          {benchmarkLoading ? (
-            <div className="h-64 flex items-center justify-center">
-              <div className="h-8 w-8 border-2 border-indigo-500 border-t-transparent rounded-full animate-spin" />
-            </div>
-          ) : benchmarks.length > 0 ? (
-            <>
-              <ResponsiveContainer width="100%" height={280}>
-                <LineChart data={benchmarkChartData}>
-                  <XAxis
-                    dataKey="date"
-                    fontSize={10}
-                    tickLine={false}
-                    axisLine={false}
-                    tickFormatter={v => {
-                      const d = new Date(v + "T00:00:00");
-                      return d.toLocaleDateString("en-CA", { month: "short", year: "2-digit" });
-                    }}
-                    interval="preserveStartEnd"
-                  />
-                  <YAxis
-                    fontSize={10}
-                    tickLine={false}
-                    axisLine={false}
-                    tickFormatter={v => `${v > 0 ? "+" : ""}${v}%`}
-                  />
-                  <Tooltip
-                    content={<GlassTooltip formatter={(v) => `${Number(v) > 0 ? "+" : ""}${Number(v).toFixed(2)}%`} />}
-                    labelFormatter={label => {
-                      const d = new Date(label + "T00:00:00");
-                      return d.toLocaleDateString("en-CA", { month: "long", day: "numeric", year: "numeric" });
-                    }}
-                  />
-                  <Legend />
-                  {benchmarks.map(b => (
-                    <Line
-                      key={b.symbol}
-                      type="monotone"
-                      dataKey={b.name}
-                      stroke={b.color}
-                      strokeWidth={2}
-                      dot={false}
-                      connectNulls
-                    />
-                  ))}
-                </LineChart>
-              </ResponsiveContainer>
-              <div className="mt-4 grid grid-cols-2 lg:grid-cols-4 gap-3">
-                {benchmarks.map(b => (
-                  <div key={b.symbol} className="flex items-center gap-2 p-2 rounded-lg bg-muted/30">
-                    <ColorDot color={b.color} className="h-2.5 w-2.5" />
-                    <div className="min-w-0 flex-1">
-                      <p className="text-xs text-muted-foreground truncate">{b.name}</p>
-                      <p className={`text-sm font-mono font-semibold ${b.returnPct >= 0 ? "text-emerald-600 dark:text-emerald-400" : "text-rose-600 dark:text-rose-400"}`}>
-                        {b.returnPct >= 0 ? "+" : ""}{b.returnPct.toFixed(2)}%
-                      </p>
-                    </div>
-                  </div>
-                ))}
-              </div>
-            </>
-          ) : (
-            <div className="h-64 flex items-center justify-center text-muted-foreground text-sm">
-              Unable to load benchmark data. Try again later.
-            </div>
-          )}
-        </CardContent>
-      </Card>}
+      {devMode && (
+        <BenchmarkChart
+          benchmarks={benchmarks}
+          benchmarkLoading={benchmarkLoading}
+          benchmarkPeriod={benchmarkPeriod}
+          setBenchmarkPeriod={setBenchmarkPeriod}
+        />
+      )}
 
       {/* ── Holdings by Account (Collapsible) ───────────────────
           Issue #25 (decision 2026-05-01): the standalone "By Holding"
@@ -1685,227 +520,12 @@ export default function PortfolioPage() {
           expand region surfaces the per-account breakdown + drill-down.
           This Holdings-by-Account panel stays as-is per the same decision
           ("the per-account button row is unchanged"). */}
-      <Card>
-        <CardHeader className="pb-2">
-          <div className="flex items-center gap-2">
-            <Wallet className="h-4 w-4 text-indigo-500" />
-            <CardTitle className="text-base">Holdings by Account</CardTitle>
-          </div>
-          <p className="text-xs text-muted-foreground">Click to expand account details</p>
-        </CardHeader>
-        <CardContent className="space-y-1">
-          {accountGroups.map(([accountName, items]) => {
-            const isExpanded = expandedAccounts.has(accountName);
-            const etfs = items.filter(h => h.assetType === "etf").length;
-            const stocks = items.filter(h => h.assetType === "stock").length;
-            const cryptos = items.filter(h => h.assetType === "crypto").length;
-            const cash = items.filter(h => h.assetType === "cash").length;
-            // Inline aggregate stats per account — sum the per-holding
-            // *Display fields the API already emits in displayCurrency.
-            const acctMktValue = items.reduce((s, h) => s + (h.marketValueDisplay ?? 0), 0);
-            const acctUnrealized = items.reduce((s, h) => s + (h.unrealizedGainDisplay ?? 0), 0);
-            // Realized + dividends are emitted in each holding's quote ccy,
-            // not display ccy — fold via marketValue/marketValueDisplay
-            // ratio (the same "implied FX" the per-row toolbar mode uses).
-            const acctRealized = items.reduce((s, h) => {
-              if (h.realizedGain == null) return s;
-              const conv = (h.marketValue && h.marketValueDisplay && h.marketValue !== 0)
-                ? h.marketValueDisplay / h.marketValue
-                : 1;
-              return s + h.realizedGain * conv;
-            }, 0);
-
-            return (
-              <div key={accountName} className="border rounded-lg overflow-hidden">
-                <button
-                  className="w-full flex items-center justify-between px-4 py-3 hover:bg-muted/30 transition-colors text-left"
-                  onClick={() => toggleAccount(accountName)}
-                >
-                  <div className="flex items-center gap-3 min-w-0">
-                    <span className="font-medium text-sm truncate">{accountName}</span>
-                    <Badge variant="outline" className="text-[10px] flex-shrink-0">
-                      {items.length} holding{items.length !== 1 ? "s" : ""}
-                    </Badge>
-                    <div className="hidden sm:flex items-center gap-1.5">
-                      {etfs > 0 && (
-                        <Badge
-                          variant="secondary"
-                          className="text-[10px] h-4"
-                          ref={(el: HTMLElement | null) => {
-                            if (el) {
-                              el.style.borderColor = ASSET_TYPE_CONFIG.etf.color;
-                              el.style.color = ASSET_TYPE_CONFIG.etf.color;
-                            }
-                          }}
-                        >
-                          {etfs} ETF
-                        </Badge>
-                      )}
-                      {stocks > 0 && (
-                        <Badge
-                          variant="secondary"
-                          className="text-[10px] h-4"
-                          ref={(el: HTMLElement | null) => {
-                            if (el) {
-                              el.style.borderColor = ASSET_TYPE_CONFIG.stock.color;
-                              el.style.color = ASSET_TYPE_CONFIG.stock.color;
-                            }
-                          }}
-                        >
-                          {stocks} Stock
-                        </Badge>
-                      )}
-                      {cryptos > 0 && (
-                        <Badge
-                          variant="secondary"
-                          className="text-[10px] h-4"
-                          ref={(el: HTMLElement | null) => {
-                            if (el) {
-                              el.style.borderColor = ASSET_TYPE_CONFIG.crypto.color;
-                              el.style.color = ASSET_TYPE_CONFIG.crypto.color;
-                            }
-                          }}
-                        >
-                          {cryptos} Crypto
-                        </Badge>
-                      )}
-                      {cash > 0 && (
-                        <Badge
-                          variant="secondary"
-                          className="text-[10px] h-4"
-                          ref={(el: HTMLElement | null) => {
-                            if (el) {
-                              el.style.borderColor = ASSET_TYPE_CONFIG.cash.color;
-                              el.style.color = ASSET_TYPE_CONFIG.cash.color;
-                            }
-                          }}
-                        >
-                          {cash} Cash
-                        </Badge>
-                      )}
-                    </div>
-                  </div>
-                  <div className="flex items-center gap-4">
-                    <div className="hidden md:flex items-center gap-3 text-xs">
-                      <div className="text-right">
-                        <p className="text-muted-foreground text-[10px]">Mkt Value</p>
-                        <p className="font-mono font-medium">{formatCurrency(acctMktValue, displayCurrency)}</p>
-                      </div>
-                      <div className="text-right">
-                        <p className="text-muted-foreground text-[10px]">Unrealized</p>
-                        <p className={`font-mono font-medium ${acctUnrealized >= 0 ? "text-emerald-600 dark:text-emerald-400" : "text-rose-600 dark:text-rose-400"}`}>
-                          {acctUnrealized >= 0 ? "+" : ""}{formatCurrency(acctUnrealized, displayCurrency)}
-                        </p>
-                      </div>
-                      <div className="text-right">
-                        <p className="text-muted-foreground text-[10px]">Realized</p>
-                        <p className={`font-mono font-medium ${acctRealized >= 0 ? "text-emerald-600 dark:text-emerald-400" : "text-rose-600 dark:text-rose-400"}`}>
-                          {acctRealized !== 0 ? `${acctRealized >= 0 ? "+" : ""}${formatCurrency(acctRealized, displayCurrency)}` : "--"}
-                        </p>
-                      </div>
-                    </div>
-                    {isExpanded ? <ChevronUp className="h-4 w-4 text-muted-foreground flex-shrink-0" /> : <ChevronDown className="h-4 w-4 text-muted-foreground flex-shrink-0" />}
-                  </div>
-                </button>
-                <AnimatePresence>
-                  {isExpanded && (
-                    <motion.div
-                      initial={{ height: 0, opacity: 0 }}
-                      animate={{ height: "auto", opacity: 1 }}
-                      exit={{ height: 0, opacity: 0 }}
-                      transition={{ duration: 0.2 }}
-                      className="overflow-hidden"
-                    >
-                      <div className="border-t">
-                        <Table>
-                          <TableHeader>
-                            <TableRow>
-                              <TableHead>Holding</TableHead>
-                              <TableHead className="text-right">Qty</TableHead>
-                              <TableHead className="text-right">Avg Cost</TableHead>
-                              <TableHead className="text-right">Price</TableHead>
-                              <TableHead className="text-right">Mkt Value</TableHead>
-                              <TableHead className="text-right">Unrealized G/L</TableHead>
-                              <TableHead className="text-right">Day Chg</TableHead>
-                              <TableHead className="text-right" />
-                            </TableRow>
-                          </TableHeader>
-                          <TableBody>
-                            {items.map(h => {
-                              const hasMetrics = h.quantity !== null && h.quantity !== 0;
-                              return (
-                                <TableRow key={h.id} className="hover:bg-muted/30 transition-colors">
-                                  <TableCell>
-                                    <div className="flex items-center gap-2">
-                                      {h.image && <img src={h.image} alt="" className="h-5 w-5 rounded-full" />}
-                                      <div>
-                                        <span className="font-medium text-sm">{h.name}</span>
-                                        {h.symbol && <Badge variant="secondary" className="ml-1 font-mono text-[10px] h-4 px-1">{h.symbol}</Badge>}
-                                        {hasMetrics && h.quantity != null && h.quantity < 0 && (
-                                          <Badge variant="outline" className="ml-1 text-[10px] h-4 px-1 border-rose-500 text-rose-600 dark:border-rose-400 dark:text-rose-400" title="Net-short position">Short</Badge>
-                                        )}
-                                      </div>
-                                    </div>
-                                  </TableCell>
-                                  <TableCell className={`text-right font-mono text-sm ${hasMetrics && h.quantity != null && h.quantity < 0 ? "text-rose-600 dark:text-rose-400" : ""}`}>
-                                    {hasMetrics && h.quantity != null
-                                      ? h.quantity.toLocaleString("en-CA", { minimumFractionDigits: 0, maximumFractionDigits: h.quantity % 1 === 0 ? 0 : 4 })
-                                      : <span className="text-muted-foreground text-xs">--</span>}
-                                  </TableCell>
-                                  <TableCell className="text-right font-mono text-sm">
-                                    {hasMetrics && h.avgCostPerShare != null
-                                      ? formatCurrency(h.avgCostPerShare, h.quoteCurrency ?? h.currency)
-                                      : <span className="text-muted-foreground text-xs">--</span>}
-                                  </TableCell>
-                                  <TableCell className="text-right font-mono text-sm">
-                                    {h.price != null ? formatCurrency(h.price, h.quoteCurrency ?? h.currency) : "--"}
-                                  </TableCell>
-                                  <TableCell className="text-right font-mono text-sm font-medium">
-                                    {hasMetrics && h.marketValueDisplay != null
-                                      ? formatCurrency(h.marketValueDisplay, displayCurrency)
-                                      : <span className="text-muted-foreground text-xs">--</span>}
-                                  </TableCell>
-                                  <TableCell className="text-right">
-                                    {hasMetrics && h.unrealizedGain != null ? (
-                                      <div>
-                                        <p className={`text-xs font-mono font-medium ${h.unrealizedGain >= 0 ? "text-emerald-600 dark:text-emerald-400" : "text-rose-600 dark:text-rose-400"}`}>
-                                          {h.unrealizedGain >= 0 ? "+" : ""}{formatCurrency(h.unrealizedGain, h.quoteCurrency ?? h.currency)}
-                                        </p>
-                                        {h.unrealizedGainPct != null && (
-                                          <p className={`text-[10px] font-mono ${h.unrealizedGainPct >= 0 ? "text-emerald-600 dark:text-emerald-400" : "text-rose-600 dark:text-rose-400"}`}>
-                                            {h.unrealizedGainPct >= 0 ? "+" : ""}{h.unrealizedGainPct.toFixed(2)}%
-                                          </p>
-                                        )}
-                                      </div>
-                                    ) : <span className="text-muted-foreground text-xs">--</span>}
-                                  </TableCell>
-                                  <TableCell className="text-right">
-                                    <DayChange pct={h.changePct} amount={h.dayChangeDisplay} currency={displayCurrency} />
-                                  </TableCell>
-                                  <TableCell className="text-right">
-                                    <Link
-                                      href={`/transactions?portfolioHolding=${encodeURIComponent(h.name)}${h.accountId ? `&accountId=${h.accountId}` : ""}`}
-                                      onClick={(e) => e.stopPropagation()}
-                                      className="text-[11px] text-primary hover:underline whitespace-nowrap"
-                                      title="View transactions for this holding in this account"
-                                    >
-                                      View txns →
-                                    </Link>
-                                  </TableCell>
-                                </TableRow>
-                              );
-                            })}
-                          </TableBody>
-                        </Table>
-                      </div>
-                    </motion.div>
-                  )}
-                </AnimatePresence>
-              </div>
-            );
-          })}
-        </CardContent>
-      </Card>
+      <HoldingsByAccount
+        accountGroups={accountGroups}
+        expandedAccounts={expandedAccounts}
+        toggleAccount={toggleAccount}
+        displayCurrency={displayCurrency}
+      />
 
       {/* Edit / create holding dialog — wraps the shared
           <HoldingEditForm> from src/components/holdings/holding-edit-form.tsx.
@@ -1944,15 +564,7 @@ export default function PortfolioPage() {
               onSave={() => {
                 setEditingHolding(null);
                 setCreatingHolding(null);
-                setLoading(true);
-                fetch(
-                  `/api/portfolio/overview?currency=${encodeURIComponent(displayCurrency)}`,
-                )
-                  .then((r) => r.json())
-                  .then((d) => {
-                    setData(d);
-                    setLoading(false);
-                  });
+                reload();
               }}
             />
           )}
@@ -1978,19 +590,4 @@ function holdingFromEnriched(h: EnrichedHolding): HoldingEditFormHolding {
     isCrypto: (h as unknown as { isCrypto?: number }).isCrypto ?? null,
     note: (h as unknown as { note?: string }).note ?? null,
   };
-}
-
-// ── Helper ──────────────────────────────────────────────────────────
-function buildBenchmarkChartData(benchmarks: BenchmarkData[]): Record<string, unknown>[] {
-  if (benchmarks.length === 0) return [];
-  const dateMap = new Map<string, Record<string, unknown>>();
-  for (const b of benchmarks) {
-    for (const pt of b.series) {
-      if (!dateMap.has(pt.date)) dateMap.set(pt.date, { date: pt.date });
-      dateMap.get(pt.date)![b.name] = pt.value;
-    }
-  }
-  return Array.from(dateMap.values()).sort((a, b) =>
-    (a.date as string).localeCompare(b.date as string)
-  );
 }
